SOLUCIONTAREAESPECIAL
Luego de 10 dias de sufrimiento, la tarea final esta completada, esta es laversion final de la tarea, todo funciona hasta aca (menos las interrupciones por transmision, no se pudieron llevar a cabo). todo lo demas funciona bien y esta organizado el documento

+++ PeripheralDrivers/Inc/SysTickDriver.h
@@ -14,7 +14,7 @@
 #define SYSTICK_LOAD_VALUE_80MHz_1ms         80000  //Numero de ciclos es 1ms
 #define SYSTICK_LOAD_VALUE_100MHz_1ms       100000  //Numero de ciclos es 1ms
 
-void config_SysTick_ms(uint8_t systemClock);
+void config_SysTick_ms(uint32_t systemClock);
 uint64_t getTicks_ms(void);
 void delay_ms(uint32_t wait_time_ms);
